requirements.txt
Flask
mysql-connector-python
flask-bcrypt
gunicorn
reportlab